Splitted request handlers to separate modules
